Reorganize code to be able to re-use functionalities in immediate run command and split the main package in smaller packages
