+++ internal/types/status.go
@@ -1,0 +1,38 @@
+package types
+
+// StatusReport contains one or more status items and is the parent object
+type StatusReport []StatusItem
+
+// StatusItem is used to serialize an individual part of the status read by the server
+type StatusItem struct {
+	Version      int    `json:"version"`
+	TimestampUTC string `json:"timestampUTC"`
+	Status       Status `json:"status"`
+}
+
+// StatusType reports the execution status
+type StatusType string
+
+const (
+	// StatusTransitioning indicates the operation has begun but not yet completed
+	StatusTransitioning StatusType = "transitioning"
+
+	// StatusError indicates the operation failed
+	StatusError StatusType = "error"
+
+	// StatusSuccess indicates the operation succeeded
+	StatusSuccess StatusType = "success"
+)
+
+// Status is used for serializing status in a manner the server understands
+type Status struct {
+	Operation        string           `json:"operation"`
+	Status           StatusType       `json:"status"`
+	FormattedMessage FormattedMessage `json:"formattedMessage"`
+}
+
+// FormattedMessage is a struct used for serializing status
+type FormattedMessage struct {
+	Lang    string `json:"lang"`
+	Message string `json:"message"`
+}
