+++ main/main.go
@@ -7,86 +7,64 @@ import (
 	"strings"
 	"time"
 
+	commands "github.com/Azure/run-command-handler-linux/internal/cmds"
+	"github.com/Azure/run-command-handler-linux/internal/constants"
 	"github.com/Azure/run-command-handler-linux/internal/handlersettings"
+	"github.com/Azure/run-command-handler-linux/internal/instanceview"
+	"github.com/Azure/run-command-handler-linux/internal/types"
+	seqnum "github.com/Azure/run-command-handler-linux/pkg/seqnumutil"
+	"github.com/Azure/run-command-handler-linux/pkg/versionutil"
 	"github.com/go-kit/kit/log"
 )
 
+// These fields are populated by govvv at compile-time.
 var (
-	// dataDir is where we store the downloaded files, logs and state for
-	// the extension handler
-	dataDir = "/var/lib/waagent/run-command-handler"
-
-	// Directory used for copying the Run Command script file to be able to RunAs a different user.
-	// It needs to copied because of permission restrictions. RunAsUser does not have permission to execute under /var/lib/waagent and its subdirectories.
-	// %s needs to be replaced by '<RunAsUser>' (RunAs username)
-	runAsDir = "/home/%s/waagent/run-command-handler-runas"
-
-	// seqNumFile holds the processed highest sequence number to make
-	// sure we do not run the command more than once for the same sequence
-	// number. Stored under dataDir.
-	//seqNumFile = "seqnum"
-
-	// most recent sequence, which was previously traced by seqNumFile. This was
-	// incorrect. The correct way is mrseq.  This file is auto-preserved by the agent.
-	mostRecentSequence = "mrseq"
-
-	// Filename where active process keeps track of process id and process start time
-	pidFilePath = "pidstart"
-
-	// downloadDir is where we store the downloaded files in the "{downloadDir}/{seqnum}/file"
-	// format and the logs as "{downloadDir}/{seqnum}/std(out|err)". Stored under dataDir
-	// multiconfig support - when extName is set we use {downloadDir}/{extName}/...
-	downloadDir = "download"
-
-	// configSequenceNumber environment variable should be set by VMAgent to sequence number
-	configSequenceNumber = "ConfigSequenceNumber"
-
-	// configExtensionName environment variable should be set by VMAgent to extension name
-	configExtensionName = "ConfigExtensionName"
-
-	configFileExtension = ".settings"
-
-	// General failed exit code when extension provisioning fails due to service errors.
-	failedExitCodeGeneral = -1
+	Version   string
+	GitCommit string
+	BuildDate string
+	GitState  string
 )
 
 func main() {
+	// After starting the program, vars from versionutil.go must be set in order to share those values across the program.
+	versionutil.Initialize(Version, GitCommit, BuildDate, GitState)
+
 	ctx := log.NewContext(log.NewSyncLogger(log.NewLogfmtLogger(
-		os.Stdout))).With("time", log.DefaultTimestamp).With("version", VersionString())
+		os.Stdout))).With("time", log.DefaultTimestamp).With("version", versionutil.VersionString())
 
 	// parse command line arguments
 	cmd := parseCmd(os.Args)
-	ctx = ctx.With("operation", strings.ToLower(cmd.name))
+	ctx = ctx.With("operation", strings.ToLower(cmd.Name))
 
 	// parse extension environment
 	hEnv, err := handlersettings.GetHandlerEnv()
 	if err != nil {
 		ctx.Log("message", "failed to parse handlerenv", "error", err)
-		os.Exit(cmd.failExitCode)
+		os.Exit(cmd.FailExitCode)
 	}
 
 	// Multiconfig support: Agent should set env variables for the extension name and sequence number
 	seqNum := -1
-	seqNumVariable := os.Getenv(configSequenceNumber)
+	seqNumVariable := os.Getenv(constants.ConfigSequenceNumber)
 	if seqNumVariable != "" {
 		seqNum, err = strconv.Atoi(seqNumVariable)
 		if err != nil {
 			ctx.Log("message", "failed to parse env variable ConfigSequenceNumber:"+seqNumVariable, "error", err)
-			os.Exit(cmd.failExitCode)
+			os.Exit(cmd.FailExitCode)
 		}
 	}
 
-	extensionName := os.Getenv(configExtensionName)
+	extensionName := os.Getenv(constants.ConfigExtensionName)
 	if extensionName != "" {
 		ctx = ctx.With("extensionName", extensionName)
-		downloadDir = downloadDir + "/" + extensionName
-		mostRecentSequence = extensionName + "." + mostRecentSequence
-		pidFilePath = extensionName + "." + pidFilePath
+		constants.DownloadDir = constants.DownloadDir + "/" + extensionName
+		constants.MostRecentSequence = extensionName + "." + constants.MostRecentSequence
+		constants.PidFilePath = extensionName + "." + constants.PidFilePath
 	}
 
 	// Read the seqNum from latest config file in case VMAgent did not set it as env variable
 	if seqNum == -1 {
-		seqNum, err = FindSequenceNumberFromConfig(hEnv.HandlerEnvironment.ConfigFolder, configFileExtension, extensionName)
+		seqNum, err = seqnum.FindSequenceNumberFromConfig(hEnv.HandlerEnvironment.ConfigFolder, constants.ConfigFileExtension, extensionName)
 		if err != nil {
 			ctx.Log("FindSequenceNumberFromConfig", "failed to find sequence number from config folder.", "error", err)
 		} else {
@@ -97,15 +75,15 @@ func main() {
 
 	// check sub-command preconditions, if any, before executing
 	ctx.Log("event", "start")
-	if cmd.pre != nil {
+	if cmd.Pre != nil {
 		ctx.Log("event", "pre-check")
-		if err := cmd.pre(ctx, hEnv, extensionName, seqNum); err != nil {
+		if err := cmd.Pre(ctx, hEnv, extensionName, seqNum); err != nil {
 			ctx.Log("event", "pre-check failed", "error", err)
-			os.Exit(cmd.failExitCode)
+			os.Exit(cmd.FailExitCode)
 		}
 	}
-	instanceView := RunCommandInstanceView{
-		ExecutionState:   Running,
+	instView := types.RunCommandInstanceView{
+		ExecutionState:   types.Running,
 		ExecutionMessage: "Execution in progress",
 		ExitCode:         0,
 		Output:           "",
@@ -114,49 +92,49 @@ func main() {
 		EndTime:          "",
 	}
 
-	reportInstanceView(ctx, hEnv, extensionName, seqNum, StatusTransitioning, cmd, &instanceView)
+	instanceview.ReportInstanceView(ctx, hEnv, extensionName, seqNum, types.StatusTransitioning, cmd, &instView)
 
 	// execute the subcommand
-	stdout, stderr, cmdInvokeError, exitCode := cmd.invoke(ctx, hEnv, &instanceView, extensionName, seqNum)
+	stdout, stderr, cmdInvokeError, exitCode := cmd.Invoke(ctx, hEnv, &instView, extensionName, seqNum)
 	if cmdInvokeError != nil {
 		ctx.Log("event", "failed to handle", "error", cmdInvokeError)
-		instanceView.ExecutionMessage = "Execution failed: " + cmdInvokeError.Error()
-		instanceView.ExecutionState = Failed
-		instanceView.EndTime = time.Now().UTC().Format(time.RFC3339)
-		instanceView.ExitCode = exitCode
-		statusToReport := StatusSuccess
+		instView.ExecutionMessage = "Execution failed: " + cmdInvokeError.Error()
+		instView.ExecutionState = types.Failed
+		instView.EndTime = time.Now().UTC().Format(time.RFC3339)
+		instView.ExitCode = exitCode
+		statusToReport := types.StatusSuccess
 
 		// If TreatFailureAsDeploymentFailure is set to true and the exit code is non-zero, set extension status to error
-		cfg, err := GetHandlerSettings(hEnv.HandlerEnvironment.ConfigFolder, extensionName, seqNum, ctx)
-		if err == nil && cfg.PublicSettings.TreatFailureAsDeploymentFailure && cmd.failExitCode != 0 {
-			statusToReport = StatusError
+		cfg, err := handlersettings.GetHandlerSettings(hEnv.HandlerEnvironment.ConfigFolder, extensionName, seqNum, ctx)
+		if err == nil && cfg.PublicSettings.TreatFailureAsDeploymentFailure && cmd.FailExitCode != 0 {
+			statusToReport = types.StatusError
 		}
 
-		reportInstanceView(ctx, hEnv, extensionName, seqNum, statusToReport, cmd, &instanceView)
-		os.Exit(cmd.failExitCode)
+		instanceview.ReportInstanceView(ctx, hEnv, extensionName, seqNum, statusToReport, cmd, &instView)
+		os.Exit(cmd.FailExitCode)
 	} else { // No error. succeeded
-		instanceView.ExecutionMessage = "Execution completed"
-		instanceView.ExecutionState = Succeeded
-		instanceView.EndTime = time.Now().UTC().Format(time.RFC3339)
-		instanceView.ExitCode = ExitCode_Okay
+		instView.ExecutionMessage = "Execution completed"
+		instView.ExecutionState = types.Succeeded
+		instView.EndTime = time.Now().UTC().Format(time.RFC3339)
+		instView.ExitCode = constants.ExitCode_Okay
 	}
 
-	instanceView.Output = stdout
-	instanceView.Error = stderr
-	reportInstanceView(ctx, hEnv, extensionName, seqNum, StatusSuccess, cmd, &instanceView)
+	instView.Output = stdout
+	instView.Error = stderr
+	instanceview.ReportInstanceView(ctx, hEnv, extensionName, seqNum, types.StatusSuccess, cmd, &instView)
 	ctx.Log("event", "end")
 }
 
 // parseCmd looks at os.Args and parses the subcommand. If it is invalid,
 // it prints the usage string and an error message and exits with code 0.
-func parseCmd(args []string) cmd {
+func parseCmd(args []string) types.Cmd {
 	if len(os.Args) != 2 {
 		printUsage(args)
 		fmt.Println("Incorrect usage.")
 		os.Exit(2)
 	}
 	op := os.Args[1]
-	cmd, ok := cmds[op]
+	cmd, ok := commands.Cmds[op]
 	if !ok {
 		printUsage(args)
 		fmt.Printf("Incorrect command: %q\n", op)
@@ -170,13 +148,13 @@ func parseCmd(args []string) cmd {
 func printUsage(args []string) {
 	fmt.Printf("Usage: %s ", os.Args[0])
 	i := 0
-	for k := range cmds {
+	for k := range commands.Cmds {
 		fmt.Print(k)
-		if i != len(cmds)-1 {
+		if i != len(commands.Cmds)-1 {
 			fmt.Printf("|")
 		}
 		i++
 	}
 	fmt.Println()
-	fmt.Println(DetailedVersionString())
+	fmt.Println(versionutil.DetailedVersionString())
 }
