--- main/status.go
@@ -1,118 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"time"
-
-	"github.com/Azure/run-command-handler-linux/internal/handlersettings"
-	"github.com/go-kit/kit/log"
-	"github.com/pkg/errors"
-)
-
-// reportStatus saves operation status to the status file for the extension
-// handler with the optional given message, if the given cmd requires reporting
-// status.
-//
-// If an error occurs reporting the status, it will be logged and returned.
-func reportStatus(ctx *log.Context, hEnv handlersettings.HandlerEnvironment, extName string, seqNum int, t StatusType, c cmd, msg string) error {
-	if !c.shouldReportStatus {
-		ctx.Log("status", "not reported for operation (by design)")
-		return nil
-	}
-	s := New(t, c.name, msg)
-	if err := Save(hEnv.HandlerEnvironment.StatusFolder, extName, seqNum, s); err != nil {
-		ctx.Log("event", "failed to save handler status", "error", err)
-		return errors.Wrap(err, "failed to save handler status")
-	}
-	return nil
-}
-
-// Save persists the status message to the specified status folder using the
-// sequence number. The operation consists of writing to a temporary file in the
-// same folder and moving it to the final destination for atomicity.
-func Save(statusFolder string, extName string, seqNo int, r StatusReport) error {
-	fn := fmt.Sprintf("%d.status", seqNo)
-	// Support multiconfig extensions where status file name should be: extName.seqNo.status
-	if extName != "" {
-		fn = extName + "." + fn
-	}
-	path := filepath.Join(statusFolder, fn)
-	tmpFile, err := ioutil.TempFile(statusFolder, fn)
-	if err != nil {
-		return fmt.Errorf("status: failed to create temporary file: %v", err)
-	}
-	tmpFile.Close()
-
-	b, err := json.MarshalIndent(r, "", "\t")
-	if err != nil {
-		return fmt.Errorf("status: failed to marshal into json: %v", err)
-	}
-
-	if err := ioutil.WriteFile(tmpFile.Name(), b, 0644); err != nil {
-		return fmt.Errorf("status: failed to path=%s error=%v", tmpFile.Name(), err)
-	}
-
-	if err := os.Rename(tmpFile.Name(), path); err != nil {
-		return fmt.Errorf("status: failed to move to path=%s error=%v", path, err)
-	}
-
-	return nil
-}
-
-// StatusReport contains one or more status items and is the parent object
-type StatusReport []StatusItem
-
-// StatusItem is used to serialize an individual part of the status read by the server
-type StatusItem struct {
-	Version      int    `json:"version"`
-	TimestampUTC string `json:"timestampUTC"`
-	Status       Status `json:"status"`
-}
-
-// StatusType reports the execution status
-type StatusType string
-
-const (
-	// StatusTransitioning indicates the operation has begun but not yet completed
-	StatusTransitioning StatusType = "transitioning"
-
-	// StatusError indicates the operation failed
-	StatusError StatusType = "error"
-
-	// StatusSuccess indicates the operation succeeded
-	StatusSuccess StatusType = "success"
-)
-
-// Status is used for serializing status in a manner the server understands
-type Status struct {
-	Operation        string           `json:"operation"`
-	Status           StatusType       `json:"status"`
-	FormattedMessage FormattedMessage `json:"formattedMessage"`
-}
-
-// FormattedMessage is a struct used for serializing status
-type FormattedMessage struct {
-	Lang    string `json:"lang"`
-	Message string `json:"message"`
-}
-
-// New creates a new Status instance
-func New(t StatusType, operation string, message string) StatusReport {
-	return []StatusItem{
-		{
-			Version:      1, // this is the protocol version do not change unless you are sure
-			TimestampUTC: time.Now().UTC().Format(time.RFC3339),
-			Status: Status{
-				Operation: operation,
-				Status:    t,
-				FormattedMessage: FormattedMessage{
-					Lang:    "en",
-					Message: message},
-			},
-		},
-	}
-}
