+++ pkg/preprocess/bom_test.go
@@ -3,15 +3,13 @@ package preprocess
 import (
 	"bytes"
 	"crypto/md5"
-	"io/ioutil"
-	"testing"
-
+	"fmt"
+	"os"
 	"path/filepath"
+	"testing"
 
 	"github.com/stretchr/testify/require"
 )
-
-import "fmt"
 
 const testDataDir = "testdata"
 
@@ -35,7 +33,7 @@ var bomTestFiles = map[string]string{
 func TestBOM_CheckTestDataIntegrity(t *testing.T) {
 	for f, sum := range bomTestFiles {
 		fp := filepath.Join(testDataDir, f)
-		b, err := ioutil.ReadFile(fp)
+		b, err := os.ReadFile(fp)
 		require.Nil(t, err, "error reading %s", fp)
 
 		hash := fmt.Sprintf("%x", md5.Sum(b))
@@ -46,7 +44,7 @@ func TestBOM_CheckTestDataIntegrity(t *testing.T) {
 func TestRemoveBOM(t *testing.T) {
 	for fn := range bomTestFiles {
 		fp := filepath.Join(testDataDir, fn)
-		b, err := ioutil.ReadFile(fp)
+		b, err := os.ReadFile(fp)
 		require.Nil(t, err, "error reading %s", fp)
 
 		n := RemoveBOM(b)
